requirements.txt
requests>=2.28.0
PyGithub>=1.59.0
click>=8.0.0
python-dotenv>=0.19.0